pyproject.toml
[project]
name = "openkoutsi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fitdecode>=0.11.0",
    "fastapi[standard]",
    "sqlalchemy[asyncio]>=2.0",
    "alembic",
    "aiosqlite",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "pydantic-settings",
    "email-validator",
]
